refactor: change index names to component name
